refactor(models): share catalog metadata overlays
Keep route-bound capabilities, context-window defaults, and catalog-owned compatibility in one existing overlay path. Consolidate the identical parameter and compatibility field merges without changing catalog behavior.

Related: #130706. Extracted as a separate cleanup from #142100.
